Corresponding to the commit on SpeckleGSA repo at around this time, the reception into GSA seems to work, but more testing needed.
