+++ SpeckleStructuralGSA/Indexer.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿/*
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -302,3 +303,4 @@ namespace SpeckleStructuralGSA
     #endregion
   }
 }
+*/
